core: fix bug caused by a serialization producing less bytes than expected
Just because we know the upper limit on the number of bytes that must be
used to extract a specific piece of information. However it's not
necessary to use all the available bytes (e.g., HP of a character can be
a small number that requires only 1 byte). So do not expect that the
serialization function is smart enough to pad the result in order to fill
all the available bytes (it could be smart enough, we have all the info,
however I'm too lazy to create a higher-order function, so the padding
can be performed by another function that already has all the necessary
info)